Added files to the framework
