+++ pom.xml
@@ -35,11 +35,13 @@
 			<artifactId>mongodb-driver-sync</artifactId>
 			<version>4.11.0</version>
 		</dependency>
+		<!-- https://mvnrepository.com/artifact/org.slf4j/slf4j-simple -->
 		<dependency>
 			<groupId>org.slf4j</groupId>
-			<artifactId>slf4j-log4j12</artifactId>
-			<version>2.0.7</version>
+			<artifactId>slf4j-simple</artifactId>
+			<version>2.0.11</version>
 		</dependency>
+
 		<!-- https://mvnrepository.com/artifact/org.apache.commons/commons-lang3 -->
 		<dependency>
 			<groupId>org.apache.commons</groupId>
