+++ src/test/java/com/playwright/demo/AnnotationDemo.java
@@ -16,13 +16,14 @@ public class AnnotationDemo {
 	static Browser browser;
 
 	// New instance for each test method:
-	BrowserContext context;
-	Page page;
+	static BrowserContext context;
+	static Page page;
 
 	@BeforeSuite
 	static void launchBrowser() {
 		playwright = Playwright.create();
 		browser = playwright.chromium().launch(new BrowserType.LaunchOptions().setHeadless(false));
+		page =  browser.newPage();
 	}
 
 	@AfterSuite
@@ -42,8 +43,10 @@ public class AnnotationDemo {
 
 	@Test
 	void shouldClickButton() {
-		page.navigate("http://www.playwright.dev");
+		page.navigate("https://playwright.dev/");
 		page.locator("[class='getStarted_Sjon']").click();
+		String currentURL = page.url();
+		System.out.println(currentURL);
 		assertEquals("Clicked", page.evaluate("result"));
 	}
 }
