+++ src/test/java/com/playwright/demo/Main.java
@@ -5,7 +5,6 @@ import com.playwright.helpers.AppConstants;
 public class Main {
 
 	public static void main(String[] args) {
-
 		System.out.printf("iTunes App ID: %d%n", AppConstants.ITUNES_APP_ID);
 		System.out.printf("iTunes HomePage Title: %s%n", AppConstants.ETSY_HOME_PAGE_TITLE);
 		System.out.printf("iTunes LoginPage Title: %s%n", AppConstants.ETSY_LOGIN_PAGE_TITLE);
